Implement filtering by any user
This commit adds a double-click behavior to user icons displayed at the
top of the board & backlog interfaces.  Double-clicking initiates a
filter much in the same way that clicking on "Assigned to me" filters by
the logged in user.

Additionally, the assigneeView & template which was previously defined
both in board & backlog, is now moved to common.